requirements.txt
numpy>=1.26.0
opencv-python>=4.0
Pillow>=10.0.0
tensorflow>=2.16.0
tf_keras>=2.16.0
scikit-learn>=1.3.0
PyDrive>=1.3.1
scipy>=1.10.0
oauth2client>=4.1.3
matplotlib>=3.7.0
protobuf>=3.20.3
imutils>=0.5.2
sympy>=1.12
click>=8.0.0
setuptools
wheel